Заменил 5 переменных одним мапом. Удалил класс money, заменил его на вышеупомянутый мап
